# frontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cadastro API",
    "description": "Dev",
    "version": "v1"
  },
  "paths": {
    "/api/pessoas": {
      "get": {
        "tags": [
          "Pessoas"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pessoa"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pessoa"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pessoa"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pessoas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pessoa"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Pessoa"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Pessoa"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              }
            }
          }
        }
      }
    },
    "/api/pessoas/{id}": {
      "get": {
        "tags": [
          "Pessoas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pessoas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pessoa"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Pessoa"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Pessoa"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Pessoas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Pessoa": {
        "required": [
          "nome"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string"
          },
          "telefone": {
            "type": "string",
            "format": "tel",
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}